requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
psycopg2-binary==2.9.10
playwright==1.48.0
pillow==11.1.0
colorthief==1.0.0
python-multipart==0.0.12
alembic==1.16.2
sqlalchemy==2.0.36
